check feature gating for wasm32
