+++ fe2o3-amqp/src/auth/scram/mod.rs
@@ -22,12 +22,10 @@ pub(crate) mod client;
 mod error;
 pub use error::*;
 
-#[cfg_attr(docsrs, doc(cfg(feature = "acceptor")))]
-#[cfg(feature = "acceptor")]
-mod server;
-
-#[cfg(feature = "acceptor")]
-pub use server::ScramAuthenticator;
+cfg_acceptor! {
+    mod server;
+    pub use server::*;
+}
 
 use crate::sasl_profile::{SCRAM_SHA_1, SCRAM_SHA_256, SCRAM_SHA_512};
 
@@ -252,68 +250,6 @@ impl ScramVersion {
         }
     }
 
-    #[cfg(feature = "acceptor")]
-    fn compute_server_first_message<'a, C>(
-        &self,
-        client_first: &'a [u8],
-        base64_server_nonce: &str,
-        credentials: &C,
-    ) -> Result<Option<ServerFirstMessage<'a>>, ServerScramErrorKind>
-    where
-        C: ScramCredentialProvider,
-    {
-        let client_first = std::str::from_utf8(client_first)?;
-
-        let client_first_message_bare = client_first
-            .strip_prefix(GS2_HEADER)
-            .ok_or(ServerScramErrorKind::CannotParseGs2Header)?;
-
-        let parts: Vec<&str> = client_first_message_bare.split(',').collect();
-
-        let username = parts
-            .first()
-            .and_then(|s| s.strip_prefix(USERNAME_KEY))
-            .ok_or(ServerScramErrorKind::CannotParseUsername)?;
-        let client_nonce = parts
-            .get(1)
-            .and_then(|s| s.strip_prefix(NONCE_KEY))
-            .ok_or(ServerScramErrorKind::CannotParseClientNonce)?;
-
-        let stored_password = match credentials.get_stored_password(username) {
-            Some(stored) => stored,
-            None => return Ok(None),
-        };
-        let base64_salt = base64::encode(stored_password.salt);
-        let iterations = stored_password.iterations.to_string();
-
-        let client_server_nonce = format!("{}{}", client_nonce, base64_server_nonce);
-
-        let mut buf = BytesMut::new();
-
-        // nonce
-        buf.put_slice(NONCE_KEY.as_bytes());
-        buf.put_slice(client_server_nonce.as_bytes());
-        buf.put_u8(b',');
-
-        // salt
-        buf.put_slice(SALT_KEY.as_bytes());
-        buf.put_slice(base64_salt.as_bytes());
-        buf.put_u8(b',');
-
-        // iterations
-        buf.put_slice(ITERATION_COUNT_KEY.as_bytes());
-        buf.put_slice(iterations.as_bytes());
-
-        Ok(Some(ServerFirstMessage {
-            username,
-            client_first_message_bare: Bytes::from(
-                client_first_message_bare.to_string().into_bytes(),
-            ),
-            client_server_nonce: Bytes::from(client_server_nonce.into_bytes()),
-            message: buf.freeze(),
-        }))
-    }
-
     fn compute_server_signature<E>(
         &self,
         salted_password: &[u8],
@@ -325,74 +261,6 @@ impl ScramVersion {
         let server_key = self.hmac(salted_password, b"Server Key")?;
         let server_signature = self.hmac(&server_key, auth_message)?;
         Ok(server_signature)
-    }
-
-    #[cfg(feature = "acceptor")]
-    fn compute_server_final_message(
-        &self,
-        client_final: &[u8],
-        client_server_nonce: &[u8],
-        client_first_message_bare: &[u8],
-        server_first_message: &[u8],
-        stored_password: &StoredPassword,
-    ) -> Result<Vec<u8>, ServerScramErrorKind> {
-        let client_final = std::str::from_utf8(client_final)?;
-        let parts: Vec<&str> = client_final.split(',').collect();
-
-        let channel_binding = parts
-            .first()
-            .and_then(|s| s.strip_prefix(CHANNEL_BINDING_KEY))
-            .ok_or(ServerScramErrorKind::CannotParseClientFinalMessage)?;
-        let channel_binding = base64::decode(channel_binding)?;
-        if channel_binding != GS2_HEADER.as_bytes() {
-            return Err(ServerScramErrorKind::InvalidChannelBinding);
-        }
-
-        let nonce = parts
-            .get(1)
-            .and_then(|s| s.strip_prefix(NONCE_KEY))
-            .ok_or(ServerScramErrorKind::CannotParseClientFinalMessage)?;
-        if nonce.as_bytes() != client_server_nonce {
-            return Err(ServerScramErrorKind::IncorrectClientFinalNonce)?;
-        }
-
-        let client_proof = parts
-            .last()
-            .and_then(|s| s.strip_prefix(PROOF_KEY))
-            .ok_or(ServerScramErrorKind::ProofNotFoundInClientFinal)?;
-
-        let without_proof_message_len =
-            client_final.len() - (client_proof.len() + PROOF_KEY.len() + 1);
-        let client_final_message_without_proof =
-            &client_final[0..without_proof_message_len].as_bytes();
-
-        let auth_message = auth_message(
-            client_first_message_bare,
-            server_first_message,
-            client_final_message_without_proof,
-        );
-
-        // Verify client proof
-        // ClientProof := ClientKey XOR ClientSignature
-        // ClientSignature := HMAC(StoredKey, AuthMessage)
-        // Inverse of XOR is XOR
-        let client_signature = self.hmac(stored_password.stored_key, &auth_message)?;
-        let client_proof = base64::decode(client_proof)?;
-        let client_key = xor(&client_proof, &client_signature)?;
-        // StoredKey := H(ClientKey)
-        let stored_key_from_client = self.h(&client_key);
-        if stored_key_from_client != stored_password.stored_key {
-            return Err(ServerScramErrorKind::AuthenticationFailed);
-        }
-
-        let server_signature_bytes = self.hmac(stored_password.server_key, &auth_message)?;
-        let server_signature = base64::encode(server_signature_bytes);
-
-        // Form server final message
-        let mut server_final = Vec::new();
-        server_final.put_slice(VERIFIER_KEY.as_bytes());
-        server_final.put_slice(server_signature.as_bytes());
-        Ok(server_final)
     }
 
     fn compute_client_proof<E>(
@@ -408,14 +276,6 @@ impl ScramVersion {
         let client_signature = self.hmac(&stored_key, auth_message)?;
         xor(&client_key, &client_signature).map_err(Into::into)
     }
-}
-
-#[cfg(feature = "acceptor")]
-pub(crate) struct ServerFirstMessage<'a> {
-    pub username: &'a str,
-    pub client_first_message_bare: Bytes,
-    pub client_server_nonce: Bytes,
-    pub message: Bytes,
 }
 
 fn client_final(client_final_message_without_proof: &[u8], client_proof: &[u8]) -> Vec<u8> {
