+++ fe2o3-amqp/src/transaction/error.rs
@@ -17,6 +17,7 @@ pub(crate) enum AllocTxnIdError {
     NotImplemented,
 
     /// Session must have dropped
+    #[cfg(not(target_arch = "wasm32"))]
     #[cfg(feature = "acceptor")]
     InvalidSessionState,
 }
@@ -25,6 +26,7 @@ pub(crate) enum AllocTxnIdError {
 #[derive(Debug)]
 pub(crate) enum DischargeError {
     /// Session must have dropped
+    #[cfg(not(target_arch = "wasm32"))]
     #[cfg(feature = "acceptor")]
     InvalidSessionState,
 
@@ -44,10 +46,12 @@ impl From<TransactionError> for DischargeError {
 #[derive(Debug)]
 pub(crate) enum CoordinatorError {
     /// The global transaction ID is not implemented yet
+    #[cfg(not(target_arch = "wasm32"))]
     #[cfg(feature = "acceptor")]
     GlobalIdNotImplemented,
 
     /// Session must have dropped
+    #[cfg(not(target_arch = "wasm32"))]
     #[cfg(feature = "acceptor")]
     InvalidSessionState,
 
@@ -64,6 +68,7 @@ impl From<AllocTxnIdError> for CoordinatorError {
     fn from(value: AllocTxnIdError) -> Self {
         match value {
             AllocTxnIdError::NotImplemented => Self::AllocTxnIdNotImplemented,
+            #[cfg(not(target_arch = "wasm32"))]
             #[cfg(feature = "acceptor")]
             AllocTxnIdError::InvalidSessionState => Self::InvalidSessionState,
         }
@@ -73,6 +78,7 @@ impl From<AllocTxnIdError> for CoordinatorError {
 impl From<DischargeError> for CoordinatorError {
     fn from(value: DischargeError) -> Self {
         match value {
+            #[cfg(not(target_arch = "wasm32"))]
             #[cfg(feature = "acceptor")]
             DischargeError::InvalidSessionState => Self::InvalidSessionState,
             DischargeError::TransactionError(error) => Self::TransactionError(error),
