+++ fe2o3-amqp/src/transaction/mod.rs
@@ -65,21 +65,12 @@ pub use owned::*;
 
 pub(crate) mod control_link_frame;
 
-#[cfg_attr(docsrs, doc(cfg(feature = "acceptor")))]
-#[cfg(feature = "acceptor")]
-pub mod coordinator;
-
-#[cfg_attr(docsrs, doc(cfg(feature = "acceptor")))]
-#[cfg(feature = "acceptor")]
-pub mod frame;
-
-#[cfg_attr(docsrs, doc(cfg(feature = "acceptor")))]
-#[cfg(feature = "acceptor")]
-pub mod manager;
-
-#[cfg_attr(docsrs, doc(cfg(feature = "acceptor")))]
-#[cfg(feature = "acceptor")]
-pub mod session;
+cfg_acceptor! {
+    pub mod coordinator;
+    pub mod frame;
+    pub mod manager;
+    pub mod session;
+}
 
 /// Trait for generics for TxnAcquisition
 #[async_trait]
