+++ fe2o3-amqp/src/util/mod.rs
@@ -3,7 +3,7 @@
 use bytes::{buf, Buf};
 use fe2o3_amqp_types::definitions::DeliveryNumber;
 use fe2o3_amqp_types::messaging::DeliveryState;
-use futures_util::{Future};
+use futures_util::Future;
 use std::io;
 use std::ops::Deref;
 use std::slice::Iter;
